third_party/WebKit/LayoutTests/shapedetection/detected-boundingBox-read-only.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/services/shape_detection/public/mojom/barcodedetection_provider.mojom.js"></script>
<script src="file:///gen/services/shape_detection/public/mojom/facedetection_provider.mojom.js"></script>
<script src="resources/big-buffer-helpers.js"></script>
<script src="resources/mock-barcodedetection.js"></script>
<script src="resources/mock-facedetection.js"></script>
<script>

const createTestForImageData = (createDetector, mock, detectionResultTest) => {
  async_test(t => {

    const img = new Image();
    img.onload = async() => {

      const canvas = document.createElement("canvas");;
      canvas.getContext("2d").drawImage(img, 0, 0);

      const detector = createDetector();
      try {
        const detectionResult = await detector.detect(canvas.getContext("2d")
            .getImageData(0, 0, canvas.width, canvas.height));
        detectionResultTest(detectionResult);
        t.done();
      } catch(error) {
        assert_unreached("Error during detect(canvas): " + error);
      }
    }

    img.src = "../media/content/greenbox.png";
  });
};

function CheckDetectedReadOnlyBoundingBoxes(detectedObject) {
  assert_readonly(detectedObject[0].boundingBox, 'x');
  assert_readonly(detectedObject[0].boundingBox, 'y');
  assert_readonly(detectedObject[0].boundingBox, 'width');
  assert_readonly(detectedObject[0].boundingBox, 'height');
  assert_readonly(detectedObject[0].boundingBox, 'top');
  assert_readonly(detectedObject[0].boundingBox, 'right');
  assert_readonly(detectedObject[0].boundingBox, 'bottom');
  assert_readonly(detectedObject[0].boundingBox, 'left');
}

// These tests verify that detectedFace's boundingBox should be DOMRectReadOnly.
generate_tests(createTestForImageData, [
  [
    "Face - detectedFace.boundingBox should be DOMRectReadOnly",
    () => { return new FaceDetector(); },
    mockFaceDetectionProvider,
    CheckDetectedReadOnlyBoundingBoxes
  ],
  [
    "Barcode - detectedBarcode.boundingBox should be DOMRectReadOnly",
    () => { return new BarcodeDetector(); },
    mockBarcodeDetectionProvider,
    CheckDetectedReadOnlyBoundingBoxes
  ]
]);

</script>
